Extract GitLab dialog presentation parts
